CSSTUDIO-3165 Move the method determineHelpLocation() from the class 'HelpBrowser' to the class 'OpenHelp'.
